requirements.txt
aioredis==1.2.0
appnope==0.1.0
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.5.1
Automat==0.7.0
backcall==0.1.0
certifi==2019.3.9
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cycler==0.10.0
daphne==2.3.0
decorator==4.4.0
Django==2.2.1
django-channels==0.7.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.9.4
future==0.17.1
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipython==7.5.0
ipython-genutils==0.2.0
iso8601==0.1.12
jedi==0.13.3
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.0
msgpack==0.6.1
numpy==1.16.3
oauthlib==3.0.1
packaging==19.0
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.2
ptyprocess==0.6.0
Pygments==2.4.0
PyHamcrest==1.9.0
pyparsing==2.4.0
pyserial==3.4
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
tornado==6.0.2
traitlets==4.3.2
Twisted==19.2.0
txaio==18.8.1
urllib3==1.25.2
wcwidth==0.1.7
zope.interface==4.6.0
